% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpct.R
\name{gpct}
\alias{gpct}
\title{Twice-Generalized Odds Ratios}
\usage{
gpct(x, explanatory_var, stratum=NULL,
        alpha=0.05, splitTies=FALSE)
}
\arguments{
\item{x}{an object of type \code{pairwise}.}

\item{explanatory_var}{a character string indicating the explanatory variable}

\item{alpha}{The acceptable type 1 error used in the test.}

\item{splitTies}{A boolean specifying how ties should be treated.
Should be \code{TRUE} for WMW Odds,
or \code{FALSE} for Agresti's GenOR.}

\item{numThreads}{The number of threads to use. If \code{numThreads<1}, the method uses that proportion of maximum available threads (rounded up).}

\item{nnt}{A boolean.
If \code{TRUE}, then print number needed to treat in addition to generalised odds ratios.}

\item{verbose}{A boolean.
If \code{TRUE}, then print both pooled odds and relative risk ratio matrices
regardless of result of statistical test.}

\item{upper}{A boolean specifying if the upper triangle
of relative risk ratios should be printed.
If \code{FALSE}, lower triangle is used instead.}
}
\value{
A list with class "\code{gpct}" containing the following:
\describe{
    \item{pooled_lnodds}{The pooled log(odds).}
    \item{pooled_lnconf.int}{(1-\code{alpha})\% Confidence intervals for pooled log(odds).}
    \item{pooled_SElnodds}{Standard error of pooled log(odds).}
    \item{pooled_SElnnull}{Standard error of pooled log(odds) under the null hypothesis.}
    \item{pooled_p}{The p-value of the test of pooled log(odds) = 1.}
    \item{pooled_rel_statistic}{Statistic of test that strata odds are equal.}
    \item{pooled_rel_p}{p-value for test that strata odds are equal.}
    \item{relative_lnodds}{A matrix giving the log of the ratio of odds between strata (generalised relative risk ratio).}
    \item{relative_selnodds}{A matrix containing the standard error of the log(relative risk ratio).}
    \item{results}{A list containing a summary of each strata measure.}
    \item{param.record}{A list containing parameters used in the test.}
}
}
\description{
Calculates Generalised Pairwise Comparisons for Trend
}
\details{
For details of Generalized Odds Ratios, see \code{\link{genodds}}.

Generalized Generalized Odds Ratios is a generalisation of Agresti's
generalized odds ratios (GenOR), capable of handling arbitrary outcomes.

This function implements Generalized Generalised Odds Ratios for
}
\references{
Our paper will go here
}
