% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inconsistencyScore.R
\name{inconsistency}
\alias{inconsistency}
\title{Inconsistency}
\usage{
inconsistency(x, numThreads=NULL, includeTies=TRUE)
}
\arguments{
\item{x}{an object of type \code{pairwise}.}

\item{numThreads}{The number of threads to use. If \code{numThreads<1}, the method uses that proportion of maximum available threads (rounded up).}
}
\description{
Calculates measures of preference inconsistency for pairwise preferences
}
