% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_definitions.R
\name{pairwise}
\alias{pairwise}
\title{Pairwise Comparison}
\usage{
pairwise(comparisons, weight=NULL,  attributes=NULL)
}
\arguments{
\item{comparisons}{A matrix containing pairwise comparison information}

\item{weight}{An optional character string giving the column name that contains the weight associated with each observation}

\item{attributes}{A data frame containing attributes for each observation}

\item{simplify}{A boolean that currently does nothing but will eventually optionally simplify data down for faster computation}
}
\value{
An object of type \code{pairwise} containing the input variables.
}
\description{
Collates variables into a \code{pairwise} object
}
