% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_definitions.R
\name{pairwise_scalar}
\alias{pairwise_scalar}
\title{Pairwise Comparison (ordered scalar)}
\usage{
pairwise_scalar(outcome, discriminant=NULL, weight=NULL,  attributes=NULL, data, simplify=FALSE, numThreads=1)
}
\arguments{
\item{outcome}{A character vector giving column names that contain outcomes.}

\item{type}{A character vector indicating if variables are \code{numeric} or \code{factor}.}

\item{decisionRule}{A list describing how winners are determined}

\item{weight}{An optional character string giving the column name that contains the weight associated with each observation}

\item{attributes}{A character vector giving column names that contain attributes}

\item{data}{A data frame, containing the specified columns}

\item{numThreads}{The number of threads to use. If \code{numThreads<1}, the method uses that proportion of maximum available threads (rounded up).}

\item{simplify}{A boolean that currently does nothing but will eventually optionally simplify data down for faster computation}
}
\value{
An object of type \code{pairwise} containing the following:
\describe{
    \item{comparisons}{A square matrix containing all-to-all comparisons of outcomes}
    \item{group}{A vector giving the groups associated with each row}
    \item{weight}{The weight of each observation}
    \item{attributes}{A data frame containing attributes for each observation}
 }
}
\description{
Calculates pairwise comparison for scalar
}
\details{
This function appears to perform best when \code{numThreads=1}.
This could be overhead outweighing benefits, or it could be something to do with linking to other functions? idk

DETAILS ABOUT DECISION DETAILS
decisionDetails should give a list of values that correspond to one of the following.

If an outcome is numeric, a single number giving the discriminating threshold to be considered a tie
If an outcome is categorical, a matrix giving the indicator value for each possible comparison
}
