% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationTest.R
\name{permutationTest}
\alias{permutationTest}
\title{Permutation Tests}
\usage{
winprop(x, splitTies=FALSE)
}
\arguments{
\item{x}{an object of type \code{pairwise}.}

\item{nSims}{a positive number giving the number of permutations to run}

\item{splitTies}{A logical vector specifying how ties should be treated.
Should be \code{TRUE} for WMW Odds,
or \code{FALSE} for Agresti's GenOR.}

\item{test}{an character vector giving the names of the tests to run \code{pairwise}.}
}
\description{
Performs permutation testing
}
