% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.survival.R
\name{simulate.survival}
\alias{simulate.survival}
\title{simulate.survival}
\usage{
simulate.survival(scales=NULL, rates=NULL, shapes=NULL, censor=NULL, heirarchy=NULL, hr, N, treatRatio = 1,
                  arrivalRate = NULL, maxt = NULL, missingAcrualRelax=NULL, minObsTime=0
                  )
}
\arguments{
\item{scales}{A vector (optionally named) containing the scale parameter for a series of weibull distributions}

\item{rates}{An alternative to scale (related by \code{rates=1/scales}).}

\item{shapes}{A vector containing the shape parameter for a series of weibull distributions}

\item{censor}{A matrix governing if certain events censor other events. See Details.}

\item{heirarchy}{An optional named vector giving the ordering of events (first event is most important).}

\item{hr}{A vector giving the hazard ratio for each event.}

\item{N}{Total sample size}

\item{treatRatio}{Ratio of treatment-to-control assignments planned.}

\item{arrivalRate}{The average arrival rate for patients (number per time unit). See details for how to estimate this.}

\item{maxt}{Maximum time to run the trial.}

\item{missingAcrualRelax}{either "N" or "maxt". Governs behaviour if the trial does not meet recruitment targets.}

\item{minObsTime}{a numeric vector used if \code{missingAcrualRelax="maxt"}. Gives the minimum amount of time that all participants are followed for.}
}
\description{
Simulate a time-to-event trial using composite survival data
}
\details{
Shape and Rate are given using the same parameterisation as pweibul, not simsurv.

This function is designed to simulate a RCT with survival endpoint data,
including realistic censoring of patients who are recruited into the trial
at a later point. As such, it is neccesary to direct the method what should happen
if the planned recruitment process fails.

This is controlled by the \code{missingAcrualRelax} parameter.
If \code{missingAcrualRelax="N"}, then the trial stops at \code{maxt} time units
regardless of current sample size.

If \code{missingAcrualRelax="maxt"}, then the trial continues until the original sample size was met,
then continues observation for an extra \code{minObsTime} time units.

In practice, this contingency should never be needed. The necessary value of
\code{arrivalRate} to accrue all patients on time can be calculated using the following command:

 \code{uniroot(f = function(x){maxt-qgamma(tol, shape=N, rate=x)},interval=c(0.1,1000))}

 Where \code{tol} is the probability of achieving the accrual target on time, and \code{maxt} and \code{N} are as described above.
}
