% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_definitions.R
\name{summary.Pairwise}
\alias{summary.Pairwise}
\title{Summarizing Pairwise Comparisons}
\usage{
summary(x, numThreads=NULL, includeTies=TRUE, countInconsistency=T, m=3)
}
\arguments{
\item{x}{an object of type \code{pairwise}.}

\item{numThreads}{The number of threads to use. If \code{numThreads<1}, the method uses that proportion of maximum available threads (rounded up).}

\item{includeTies}{This really needs to be deleted in the final version}

\item{m}{The length of cycle for measuring inconsistency}
}
\description{
Produces a summary of a \code{Pairwise} object.
}
\details{
Details here
}
