% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winProportion.R
\name{winProportion}
\alias{winProportion}
\title{Multi-group Win Proportion}
\usage{
winprop(x, splitTies=FALSE)
}
\arguments{
\item{x}{an object of type \code{pairwise}.}

\item{splitTies}{A boolean specifying how ties should be treated.
Should be \code{TRUE} for WMW Odds,
or \code{FALSE} for Agresti's GenOR.}

\item{alpha}{The acceptable type 1 error used in the test.}
}
\description{
Calculates the win proportion for multiple groups
}
\references{
Our paper will go here
}
