% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winRatio.R
\name{winRatio}
\alias{winRatio}
\title{Win Ratio for Composite Endpoints}
\usage{
winRatio <- function(x, group, stratum=NULL,
                    pool_method = c("unweighted","inversevariance","size"),
                    alpha=0.05, splitTies=FALSE, numThreads=NULL)
}
\arguments{
\item{x}{an object of type \code{pairwise}.}

\item{group}{a character string indicating the group variable}

\item{stratum}{A character vector indicating which covariates should be used for stratification}

\item{pool_method}{A character vector indicating how each stratum should be weighted}

\item{alpha}{The acceptable type 1 error used in the test.}

\item{splitTies}{A boolean specifying how ties should be treated.
Should be \code{TRUE} for Win Odds,
or \code{FALSE} for Win Ratio.}

\item{numThreads}{The number of threads to use. If \code{numThreads<1}, the method uses that proportion of maximum available threads (rounded up).}
}
\value{
A list containing the following:
\describe{
    \item{pooled_lnodds}{The pooled log(odds).}
    \item{pooled_lnconf.int}{(1-\code{alpha})\% Confidence intervals for pooled log(odds).}
    \item{pooled_SElnodds}{Standard error of pooled log(odds).}
    \item{pooled_SElnnull}{Standard error of pooled log(odds) under the null hypothesis.}
    \item{pooled_p}{The p-value of the test of pooled log(odds) = 1.}
    \item{pooled_rel_statistic}{Statistic of test that strata odds are equal.}
    \item{pooled_rel_p}{p-value for test that strata odds are equal.}
    \item{relative_lnodds}{A matrix giving the log of the ratio of odds between strata (generalised relative risk ratio).}
    \item{relative_selnodds}{A matrix containing the standard error of the log(relative risk ratio).}
    \item{results}{A list containing a summary of each strata measure.}
    \item{param.record}{A list containing parameters used in the test.}
}
}
\description{
Calculates the (optionally stratified) Win Ratio for Composite Endpoints.
}
\details{
Need to write this
}
\references{
Dong, stratified win ratio
Dong, generalized analytic etc etc
}
